src/main/java/com/anzymus/neogeo/hiscores/service/RssService.java
/**
 *     Copyright (C) 2011 Julien SMADJA <julien dot smadja at gmail dot com>
 *
 *     Licensed under the Apache License, Version 2.0 (the "License");
 *     you may not use this file except in compliance with the License.
 *     You may obtain a copy of the License at
 *
 *             http://www.apache.org/licenses/LICENSE-2.0
 *
 *     Unless required by applicable law or agreed to in writing, software
 *     distributed under the License is distributed on an "AS IS" BASIS,
 *     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *     See the License for the specific language governing permissions and
 *     limitations under the License.
 */

package com.anzymus.neogeo.hiscores.service;

import javax.ejb.EJB;
import javax.ejb.Stateless;
import javax.ejb.TransactionAttribute;
import javax.ejb.TransactionAttributeType;
import com.anzymus.neogeo.hiscores.domain.Timeline;
import com.anzymus.neogeo.hiscores.service.TimelineService;
import com.anzymus.neogeo.hiscores.service.rss.TimelineRss;

@Stateless
@TransactionAttribute(TransactionAttributeType.SUPPORTS)
public class RssService {

    @EJB
    TimelineService timelineService;

    public String getTimeline() {
        Timeline timeline = timelineService.createTimeline();
        return new TimelineRss(timeline).toString();
    }

}
